CCTVLIVE/cctv-ai-cv/requirements.txt
antlr4-python3-runtime==4.9.3
bidict==0.23.1
blinker==1.7.0
cachetools==5.3.3
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
cmake==3.28.3
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
easydict==1.10
exceptiongroup==1.1.3
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.12.2
filterpy==1.4.5
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.49.0
future==0.18.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.57.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.3
importlib-resources==6.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.4
lit==17.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
minio==7.1.16
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.2
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.0.9
pyrabbit==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.0
python-socketio==5.9.0
pynvml
pytz==2023.3
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.30.0
six==1.16.0
supervisor==4.2.5
sympy==1.12
tifffile==2024.2.12
torch
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.185
urllib3==2.0.3
wcwidth==0.2.13
Werkzeug==3.0.1
yt-dlp==2024.11.4
zipp==3.16.2
zmq==0.0.0
